BicycleFrontEnd/src/app/login/signup/signup.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { SignupComponent } from './signup.component';

describe('SignupComponent', () => {
  let component: SignupComponent;
  let fixture: ComponentFixture<SignupComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [SignupComponent]
    })
      .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(SignupComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });
  it('form is invalaid when empty', () => {
    expect(component.formModel.valid).toBeFalsy();
  });
  it('email field validity', () => {
    let email = component.formModel.controls['email'];
    let errors = email.errors || {}
    expect(errors['required']).toBeTruthy();
  });
});
